+++ tests/test_e2e.py
@@ -87,7 +87,7 @@ def test_e2e(
             insert(db, commit=False)
             r = cli_runner.invoke(["diff"])
             assert r.exit_code == 0
-            assert re.match(fr"\+\+\+ {table}:\w+=\d+$", r.stdout.splitlines()[0])
+            assert re.match(fr"\+\+\+ {table}:feature:\d+$", r.stdout.splitlines()[0])
 
             # commit it
             r = cli_runner.invoke(["commit", "-m", "commit-1"])